+++ NEWS.md
@@ -1,3 +1,9 @@
+# git2rdata 0.4.1
+
+* Add `update_metadata()` to update the description of a `git2rdata` object.
+  See `vignette("metadata")` for more details.
+* Update the checklist and pkgdown infrastructure.
+
 # git2rdata 0.4.0
 
 ## New features